apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
redis==5.0.4
celery==5.4.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
stripe==9.9.0
boto3==1.34.110
spacy==3.7.4
langdetect==1.0.9
python-dotenv==1.0.1
structlog==24.1.0
prometheus-client==0.20.0
pytest==8.2.1
pytest-asyncio==0.23.7
